Made the ProgressBar more efficient.
Before the ProgressBar would continuously check for changes to the
`activeRequests` variable, by `sleeping` and checking in a loop.

Now the `progressBarFetch` informs the `ProgressBar` that a request
was sent, and the ProgressBar only starts the `flow` at that particular
time.

The `flow` now has a new mode called 'hibernate' which represents the
state in which the ProgressBar is waiting for the `progressBarFetch` to
tell if a request has been sent. Once the animation has been completed
it moves back to hibernation.

When the mode is 'init' it will move either to 'active' or 'hibernate'.
It will move to 'active' if there are requests within '100' milliseconds,
if there are no requests, because they were completed really fast, it
moves back to 'hibernate' mode, because we do not want to show the
animation needlessly.